requirements.txt
allure-pytest
allure-python-commons
Appium-Python-Client
attrs
axe-selenium-python
certifi
cffi
charset-normalizer
colorama
h11
idna
iniconfig
Jinja2
MarkupSafe
outcome
packaging
pip
pluggy
PyAutoGUI
pycparser
PySocks
pytest
pytest-html
pytest-metadata
pytest-ordering
requests
selenium
setuptools
sniffio
sortedcontainers
trio
trio-websocket
urllib3
wsproto
Pillow